// frontend/src/assets/ts/types.ts
export interface Userhurt {
    id: string;
    user_id: string;
    fill_time: string;
    neck: number;
    right_upper_arm: number;
    right_shoulder: number;
    right_lower_arm: number;
    right_hand: number;
    left_upper_arm: number;
    left_shoulder: number;
    left_lower_arm: number;
    left_lower_leg: number;
    left_hand: number;
    left_upper_leg: number;
    left_ankle: number;
    left_feet: number;
    left_knee: number;
    right_lower_leg: number;
    right_ankle: number;
    right_upper_leg: number;
    right_feet: number;
    right_knee: number;
    abdomen: number;
    lower_body: number;
    upper_body: number;
    right_ear: number;
    left_ear: number;
    head: number;
    right_eye: number;
    mouth: number;
    left_eye: number;
    nose: number;
    back_head: number;
    back_neck: number;
    left_elbow: number;
    right_elbow: number;
    lower_back: number;
    back: number;
    butt: number;
    right_upper_shoulder: number;
}

export interface Usertime {
    id: string;
    user_id: string;
    fill_time: string;
    neck: boolean;
    right_upper_arm: boolean;
    right_shoulder: boolean;
    right_lower_arm: boolean;
    right_hand: boolean;
    left_upper_arm: boolean;
    left_shoulder: boolean;
    left_lower_arm: boolean;
    left_lower_leg: boolean;
    left_hand: boolean;
    left_upper_leg: boolean;
    left_ankle: boolean;
    left_feet: boolean;
    left_knee: boolean;
    right_lower_leg: boolean;
    right_ankle: boolean;
    right_upper_leg: boolean;
    right_feet: boolean;
    right_knee: boolean;
    abdomen: boolean;
    lower_body: boolean;
    upper_body: boolean;
    right_ear: boolean;
    left_ear: boolean;
    head: boolean;
    right_eye: boolean;
    mouth: boolean;
    left_eye: boolean;
    nose: boolean;
    back_head: boolean;
    back_neck: boolean;
    left_elbow: boolean;
    right_elbow: boolean;
    lower_back: boolean;
    back: boolean;
    butt: boolean;
    right_upper_shoulder: boolean;
}